+++ packages/client-wallets/RELEASES.md
@@ -79,3 +79,13 @@ Based on:
 - [typescript v1.0.0-alpha.9] packages/client-wallets
 ### Releases
 - [NPM v1.0.0-alpha.9] https://www.npmjs.com/package/@formance/sdk-wallets/v/1.0.0-alpha.9 - packages/client-wallets
+
+## 2024-08-06 00:11:34
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.352.2 (2.385.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v1.0.0-alpha.10] packages/client-wallets
+### Releases
+- [NPM v1.0.0-alpha.10] https://www.npmjs.com/package/@formance/sdk-wallets/v/1.0.0-alpha.10 - packages/client-wallets